move models to model folder
